hard/python-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
pydantic[email]>=2.10.0
PyJWT>=2.9.0
pytest>=8.0.0
